+++ superuser_plugin_unix/src/superuser_plugin_unix.h
@@ -5,10 +5,19 @@
 
 #define FFI_PLUGIN_EXPORT
 
+typedef unsigned int ERRCODE;
+
 // Determine user who execute this program is root.
 FFI_PLUGIN_EXPORT bool is_root();
 
+// Determine user is a group member, which eligable to execute program as root by calling
+// sudo command.
+//
+// This method requires sudo bundled in OS already. Normally, majority of UNIX or liked
+// system.
+FFI_PLUGIN_EXPORT ERRCODE is_sudo_group(bool* result);
+
 // Obtain name of user.
-FFI_PLUGIN_EXPORT int get_uname(char** result);
+FFI_PLUGIN_EXPORT ERRCODE get_uname(char** result);
 
 #endif